+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/rest/user/get/RestGetUserAction.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.identity.rest.user.read;
+package org.opensearch.identity.rest.user.get;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonNode;
